+++ my-tiny-spring/src/main/java/com/yrb/tinyioc/aop/AopProxy.java
@@ -1,0 +1,11 @@
+package com.yrb.tinyioc.aop;
+
+/**
+ * @author bjyangrubing
+ * @createTime 2017/8/7 20:20
+ * Description: AOP代理
+ */
+public interface AopProxy
+{
+	Object getProxy();
+}