+++ my-tiny-spring/src/main/java/com/yrb/tinyioc/beans/BeanDefinition.java
@@ -13,6 +13,10 @@ public class BeanDefinition
 
 	private String beanClassName;
 
+	public BeanDefinition()
+	{
+	}
+
 	public PropertyValues getPropertyValues()
 	{
 		return propertyValues;